Add Rate coworkingSpace

+++ CoworkingReservation.Domain/DTOs/CoworkingSpaceListItemDTO.cs
@@ -12,7 +12,7 @@ namespace CoworkingReservation.Domain.DTOs
         public string Name { get; set; }
         public AddressDTO Address { get; set; }
         public PhotoResponseDTO? CoverPhoto { get; set; }
-        public int Rate { get; set; }
+        public float Rate { get; set; }
         public decimal PricePerDay { get; set; }
     }
 }
